Specific_variable form added successfully.
this form is added in the previous sectionc locaiton in the menu_patient.xml.

However, other unnecessary files are not deleted yet. isdelete variable is switched off in the code though there has variable in the database table.

Customized sqlite syntax for this table is given bellow-
CREATE TABLE SPECIFICVAR (
    PatientID  VARCHAR( 20 ),
    FacilityID VARCHAR( 5 ),
    CC2        VARCHAR( 2 ),
    CC3        VARCHAR( 1 ),
    CC5        VARCHAR( 2 ),
    CC6        VARCHAR( 1 ),
    CC8        VARCHAR( 3 ),
    CC9        VARCHAR( 3 ),
    CC10       VARCHAR( 1 ),
    CC11       VARCHAR( 1 ),
    isdelete   INT,
    StartTime  VARCHAR( 5 ),
    EndTime    VARCHAR( 5 ),
    DeviceID   VARCHAR( 10 ),
    EntryUser  VARCHAR( 10 ),
    Lat        VARCHAR( 20 ),
    Lon        VARCHAR( 20 ),
    EnDt       DATETIME,
    Upload     INT,
    modifyDate DATETIME,
    CONSTRAINT pk_SPECIFICVAR PRIMARY KEY ( PatientID, FacilityID )
);